Address code review comment
