+++ src/ui/components/IconPicker/iconPickerIcons.ts
@@ -1,0 +1,5 @@
+import { HOUSEHOLD_ICON_IMAGES, type HouseholdIconKey } from "../IconV2";
+
+export const ICONS_LIBRARY = Object.keys(
+  HOUSEHOLD_ICON_IMAGES,
+) as HouseholdIconKey[];
