+++ src/ui/components/IconPicker/index.ts
@@ -1,0 +1,2 @@
+export type { IconPickerHandle } from "./IconPicker";
+export { default } from "./IconPicker";
